+++ src/common/AddressFilter/AddressFilter.js
@@ -1,10 +1,16 @@
-import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
+import { useMemo } from 'react';
 
 import { SelectionFilter } from '@folio/stripes-acq-components';
+
 import { getAddressOptions } from '../utils';
 
-const AddressFilter = ({ addresses, ...rest }) => {
+const DEFAULT_ADDRESSES = [];
+
+const AddressFilter = ({
+  addresses = DEFAULT_ADDRESSES,
+  ...rest
+}) => {
   const options = useMemo(() => getAddressOptions(addresses), [addresses]);
 
   return (
@@ -17,10 +23,6 @@ const AddressFilter = ({ addresses, ...rest }) => {
 
 AddressFilter.propTypes = {
   addresses: PropTypes.arrayOf(PropTypes.object),
-};
-
-AddressFilter.defaultProps = {
-  addresses: [],
 };
 
 export default AddressFilter;